Module asciii::actions

General actions

Modules

error

Structs

ApiKeys
ApiKeys store

Dues

MetaStore
Shared extra information stored in the repo

Functions

archive_all_projects

archive_projects

calendar
Produces a calendar from the selected StorageDir

calendar_and_tasks
Command CALENDAR

calendar_with_tasks

clone_remote
Clone the repo

csv

delete_project_confirmation

dues
Command DUES

get_api_keys
get ApiKeys for server

parse_meta
Parses meta store

projects_to_csv
Produces a csv string from a list of Projects

spec
Testing only, tries to run complete spec on all projects. TODO make this not panic :D TODO move this to spec::all_the_things

store_meta

unarchive_projects
Command UNARCHIVE TODO: return a list of files that have to be updated in git

with_projects
Helper method that passes projects matching the search_terms to the passt closure f