Struct asciii::config::ConfigReader

pub struct ConfigReader {
    pub path: PathBuf,
    // some fields omitted
}

Looks for a configuration yaml in your HOME_DIR

Fields
path: PathBuf

Path of config file

Methods

impl ConfigReader

fn path_home() -> PathBuf

The Path of the config file.

fn new() -> Result<ConfigReader, YamlError>

Opens config from self.path() and parses Yaml right away.

fn get(&self, path: &str) -> Option<&Yaml>

Returns whatever it finds in that position

Supports simple path syntax: top/middle/child/node

fn get_char(&self, key: &str) -> Option<char>

Returns the first character.

Panics

This panics if nothing is found. You should have a default config for everything that you use.

fn get_str(&self, key: &str) -> Option<&str>

Returns the string in the position or an empty string

fn get_strs(&self, key: &str) -> Option<Vec<&str>>

Returns the a vec of &strs if possible

fn get_to_string(&self, key: &str) -> Option<String>

Returns the string in the position or an empty string

Panics

This panics if nothing is found. You should have a default config for everything that you use.

fn get_f64(&self, key: &str) -> Option<f64>

Tries to get the config field as float

fn get_bool(&self, key: &str) -> bool

Returns the boolean in the position or false

Panics

This panics if nothing is found. You should have a default config for everything that you use.