Function asciii::fill_docs::fill_template

pub fn fill_template<E: Exportable>(document: E, template_file: &Path) -> PathBuf

Takes an exportable and a template path and does it's thing.

Returns path to created file, potenially in a tempdir.