Enum asciii::project::ComputedField

pub enum ComputedField {
    Responsible,
    OfferNumber,
    InvoiceNumber,
    InvoiceNumberLong,
    Name,
    Final,
    Age,
    OurBad,
    TheirBad,
    Year,
    Caterers,
    ClientFullName,
    Wages,
    SortIndex,
    Date,
    Invalid,
}

Project::get() allows accessing fields within the raw yaml data structure. Computed fields are fields that are not present in the document but computed.

ComputedFields is an automatically generated type that allows iterating of the variants of this Enum.

Variants
Responsible

Usually storage, or in legacy part of signature

OfferNumberInvoiceNumber

Pretty version of invoice/number: "R042"

InvoiceNumberLong

Pretty version of invoice/number including year: "R2016-042"

Name

Overall Cost Project, including taxes

FinalAgeOurBadTheirBadYearCaterersClientFullNameWagesSortIndex

Sorting index

DateInvalid
Methods

impl ComputedField

fn iter_variants() -> ComputedFields

impl ComputedField

fn iter_variant_names() -> ComputedFieldNames

impl ComputedField

fn get(&self, project: &Project) -> Option<String>