Module asciii::project::spec

Implements the ascii invoicer project file specification.

This does all of the heavy lifting. The implementation is separated into sub-modules which take care of separate objectives. Most of the functions in these modules take the yaml data directly as reference. Each module contains a validate() function which ought to be kept up to date.

Structs

Employee
Represents an Employee

Event
Describes either the coarse begin and end date of the event or holds a list of distinct EventTimes

EventTime
Holds the time of the beginning and end of an event

Traits

HasEmployees
Something that has employees

HasEvents
Something that has events

Invoicable
Stage 2: requirements for an invoice

IsClient
Everything about the client

IsProject
Stage 0: the Project itself

IsProjectExt
Extended functionality for projects

Offerable
Stage 1: requirements for an offer

Redeemable
Stage 3: when an IsProject is redeem and can be archived

Validatable
Every other trait in this module ought to be Validatable

Functions

print_specresult