Enum asciii::project::spec::VirtualField

pub enum VirtualField {
    Responsible,
    OfferNumber,
    InvoiceNumber,
    InvoiceNumberLong,
    Name,
    Final,
    Age,
    Year,
    Caterers,
    ClientFullName,
    Invalid,
}

Project::get() allows accessing fields within the raw yaml data structure. Virtual fields are fields that are not present in the document but computed.

VirtualFields is an automatically generated type that allows iterating of the variants of this Enum.

Variants
Responsible

Usually storage, or in legacy part of signature

OfferNumber

Pretty version of invoice/number: "R042"

InvoiceNumberInvoiceNumberLong

Pretty version of invoice/number including year: "R2016-042"

Name

Overall Cost Project, including taxes

FinalAgeYearCaterersClientFullNameInvalid
Methods

impl VirtualField

fn iter_variants() -> VirtualFields

impl VirtualField

fn iter_variant_names() -> VirtualFieldNames

impl VirtualField

fn get(&self, project: &Project) -> Option<String>