Struct asciii::project::Project

pub struct Project { /* fields omitted */ }

Represents a Project.

A project is storable, contains products, and you can create an offer or invoice from it. The main implementation is done in spec.

Methods

impl Project

pub fn yaml(&self) -> &Yaml

Access to inner data

pub fn open<S: AsRef<OsStr> + ?Sized>(pathish: &S) -> Result<Project, Error>

Opens a project from file path;

pub fn parse_yaml(&self) -> Result<Project, Error>

import from yaml file

pub fn dump_yaml(&self) -> String

pub fn to_json(&self) -> Result<String, Error>

export to JSON

pub fn from_file_content(content: &str) -> Result<Project, Error>

Used mostly for testing purposes

pub fn field(&self, path: &str) -> Option<String>

wrapper around yaml::get() with replacement

pub fn client(&self) -> Client

Returns the struct Client, which abstracts away client specific stuff.

pub fn offer(&self) -> Offer

Returns the struct Offer, which abstracts away offer specific stuff.

pub fn invoice(&self) -> Invoice

Returns the struct Invoice, which abstracts away invoice specific stuff.

pub fn hours(&self) -> Hours

Returns the struct Invoice, which abstracts away invoice specific stuff.

pub fn is_ready_for_offer(&self) -> SpecResult

Ready to produce offer.

Ready to send an offer to the client.

pub fn is_ready_for_invoice(&self) -> SpecResult

Valid to produce invoice

Ready to send an invoice to the client.

pub fn is_ready_for_archive(&self) -> SpecResult

Completely done and in the past.

Ready to be **h:

pub fn to_csv(&self, bill_type: BillType) -> Result<String, Error>

pub fn debug(&self) -> Debug

pub fn empty_fields(&self) -> Vec<String>

Check Templated for replacable markers

pub fn replace_field(&self, field: &str, value: &str) -> Result<(), Error>

Fill certain field

pub fn our_bad(&self) -> Option<Duration>

Time between event and creation of invoice

pub fn their_bad(&self) -> Option<Duration>

Time between creation of invoice and payment

pub fn to_tasks(&self) -> Calendar

What I need to do

Produces an iCal calendar from this project.