Module asciii::storage

Manages file structure of templates, working directory and archives.

This module takes care of project file management.

Your ordinary file structure would look something like this:

# root dir
├── working
│   └── Project1
│       └── Project1.yml
├── archive
│   ├── 2013
│   └── 2014
│       └── R036_Project3
│           ├── Project3.yml
│           └── R036 Project3 2014-10-08.tex
...

Re-exports

pub use self::error::StorageError;

pub use self::storable::*;

Modules

error
Error that may occur in Storage

repo

storable
Contains the Storable trait that storable projects must implement.

Structs

Paths

ProjectList
Wrapper around Vec<Storable>

Projects
Container keeping all opened projects.

Storage
Manages project file storage.

Enums

StorageDir
Used to identify what directory you are talking about.

StorageSelection
A description from which we can open Storables

Functions

get_storage_path
Interprets storage path from config.

list_path_content
Basically ls, returns a list of paths.

setup
Sets up an instance of Storage.

setup_with_git
Sets up an instance of Storage, with git turned on.

Type Definitions

ProjectsByYear

Year
Year = i32