Trait asciii::storage::traits::Storing

pub trait Storing<L: Storable>: Sized {
    fn new<P: AsRef<Path>>(root: P, working: &str, archive: &str, template: &str) -> StorageResult<Self>;
    fn root_dir(&self) -> &Path;
    fn working_dir(&self) -> &Path;
    fn archive_dir(&self) -> &Path;
    fn templates_dir(&self) -> &Path;

    fn list_path_content(&self, path: &Path) -> StorageResult<Vec<PathBuf>> { ... }
    fn create_dirs(&self) -> StorageResult<()> { ... }
    fn create_archive(&self, year: Year) -> StorageResult<PathBuf> { ... }
    fn list_template_files(&self) -> StorageResult<Vec<PathBuf>> { ... }
    fn list_template_names(&self) -> StorageResult<Vec<String>> { ... }
    fn get_template_file(&self, name: &str) -> StorageResult<PathBuf> { ... }
    fn list_archives(&self) -> StorageResult<Vec<PathBuf>> { ... }
    fn list_years(&self) -> StorageResult<Vec<Year>> { ... }
    fn create_project(&self, project_name: &str, template_name: &str) -> StorageResult<L> { ... }
    fn archive_project_by_name(&self, name: &str, year: Year, prefix: Option<String>) -> StorageResult<PathBuf> { ... }
    fn archive_project(&self, project: &L, year: Year) -> StorageResult<PathBuf> { ... }
    fn unarchive_project(&self, project: &L) -> StorageResult<PathBuf> { ... }
    fn unarchive_project_file(&self, archived_file: &Path) -> StorageResult<PathBuf> { ... }
    fn search_projects(&self, dir: StorageDir, search_term: &str) -> StorageResult<Vec<PathBuf>> { ... }
    fn search_multiple_projects(&self, dir: StorageDir, search_terms: &[&str]) -> StorageResult<Vec<PathBuf>> { ... }
    fn get_project_dir(&self, name: &str, directory: StorageDir) -> StorageResult<PathBuf> { ... }
    fn get_project_file(&self, directory: &Path) -> StorageResult<PathBuf> { ... }
    fn get_project_name(&self, directory: &Path) -> StorageResult<String> { ... }
    fn get_project_dir_from_archive(&self, name: &str, year: Year) -> StorageResult<PathBuf> { ... }
    fn list_project_folders(&self, directory: StorageDir) -> StorageResult<Vec<PathBuf>> { ... }
    fn list_empty_project_dirs(&self, directory: StorageDir) -> StorageResult<Vec<PathBuf>> { ... }
    fn list_project_files(&self, directory: StorageDir) -> StorageResult<Vec<PathBuf>> { ... }
    fn filter_project_files<F>(&self, directory: StorageDir, filter: F) -> StorageResult<Vec<PathBuf>> where F: FnMut(&PathBuf) -> bool { ... }
    fn open_project_files(&self, dir: StorageDir) -> StorageResult<ProjectList<L>> { ... }
}

Base functionality for archiving and unarchiving etc

Required Methods

fn new<P: AsRef<Path>>(root: P, working: &str, archive: &str, template: &str) -> StorageResult<Self>

fn root_dir(&self) -> &Path

Getter path to root directory.

fn working_dir(&self) -> &Path

Getter path to working directory.

fn archive_dir(&self) -> &Path

Getter path to archive directory.

fn templates_dir(&self) -> &Path

Getter path to templates directory.

Provided Methods

fn list_path_content(&self, path: &Path) -> StorageResult<Vec<PathBuf>>

Generic Filesystem wrapper.

fn create_dirs(&self) -> StorageResult<()>

Creates the basic dir structure inside the storage directory.

└── root
    ├── archive
    ├── templates
    └── working

If the directories already exist as expected, that's fine TODO ought to fail when storage_dir already contains directories that do not correspond with the names given in this setup.

fn create_archive(&self, year: Year) -> StorageResult<PathBuf>

Creates an archive for a certain year. This is a subdirectory under the archive directory.

└── root
    ├── archive
        ├── 2001
    ...

fn list_template_files(&self) -> StorageResult<Vec<PathBuf>>

Produces a list of files in the template_dir()

fn list_template_names(&self) -> StorageResult<Vec<String>>

Produces a list of names of all template filess in the templates_dir()

fn get_template_file(&self, name: &str) -> StorageResult<PathBuf>

Returns the Path to the template file by the given name, maybe.

fn list_archives(&self) -> StorageResult<Vec<PathBuf>>

Produces a list of paths to all archives in the archive_dir. An archive itself is a folder that contains project dirs, therefore it essentially has the same structure as the working_dir, with the difference, that the project folders may be prefixed with the projects index, e.g. an invoice number etc.

fn list_years(&self) -> StorageResult<Vec<Year>>

Produces a list of years for which there is an archive.

fn create_project(&self, project_name: &str, template_name: &str) -> StorageResult<L>

Takes a template file and stores it in the working directory, in a new project directory according to it's name.

fn archive_project_by_name(&self, name: &str, year: Year, prefix: Option<String>) -> StorageResult<PathBuf>

Moves a project folder from /working dir to /archive/$year.

fn archive_project(&self, project: &L, year: Year) -> StorageResult<PathBuf>

Moves a project folder from /working dir to /archive/$year. Also adds the project.prefix() to the folder name.

└── root
    ├── archive
        ├── 2001
            ├── R0815_Birthdayparty
    ...

fn unarchive_project(&self, project: &L) -> StorageResult<PathBuf>

Moves a project folder from /working dir to /archive/$year.

fn unarchive_project_file(&self, archived_file: &Path) -> StorageResult<PathBuf>

Moves a project folder from /working dir to /archive/$year.

fn search_projects(&self, dir: StorageDir, search_term: &str) -> StorageResult<Vec<PathBuf>>

Matches StorageDir's content against a term and returns matching project files.

fn search_multiple_projects(&self, dir: StorageDir, search_terms: &[&str]) -> StorageResult<Vec<PathBuf>>

Matches StorageDir's content against a term and returns matching project files. TODO add search_multiple_projects_deep

fn get_project_dir(&self, name: &str, directory: StorageDir) -> StorageResult<PathBuf>

Tries to find a concrete Project.

fn get_project_file(&self, directory: &Path) -> StorageResult<PathBuf>

Locates the project file inside a folder.

This is the first file with the super::PROJECT_FILE_EXTENSION in the folder

fn get_project_name(&self, directory: &Path) -> StorageResult<String>

fn get_project_dir_from_archive(&self, name: &str, year: Year) -> StorageResult<PathBuf>

fn list_project_folders(&self, directory: StorageDir) -> StorageResult<Vec<PathBuf>>

Produces a list of project folders.

fn list_empty_project_dirs(&self, directory: StorageDir) -> StorageResult<Vec<PathBuf>>

Produces a list of empty project folders.

fn list_project_files(&self, directory: StorageDir) -> StorageResult<Vec<PathBuf>>

Produces a list of project files.

fn filter_project_files<F>(&self, directory: StorageDir, filter: F) -> StorageResult<Vec<PathBuf>> where F: FnMut(&PathBuf) -> bool

fn open_project_files(&self, dir: StorageDir) -> StorageResult<ProjectList<L>>

Behaves like list_project_files() but also opens projects directly.