Trait asciii::templater::IsKeyword

pub trait IsKeyword {
    fn is_keyword(&self) -> bool;
    fn get_keyword(&self) -> Option<String>;
    fn list_keywords(&self) -> Vec<String>;
    fn map_keywords<F>(&self, closure: F) -> String where F: Fn(&str) -> String;
}

Simple template style keyword replacement.

This allows replacing a known set of keywords looking like __THIS__. Here it is implemented for Deref<Target=str>.

Required Methods

fn is_keyword(&self) -> bool

Checks if the whole string is a keyword

fn get_keyword(&self) -> Option<String>

Captures keywords from string.

fn list_keywords(&self) -> Vec<String>

Well, it lists the keywords in a string, duh!

fn map_keywords<F>(&self, closure: F) -> String where F: Fn(&str) -> String

This one is usefull.

Takes a clorsure that replaces keywords. Careful, this replaces either way! If you get a keywords you don't want to replace, please place it back where you got it from.

Example

.map_keywords|keyword| match data.get(keyword){
    Some(content) => String::from(*content),
    None => format!("__{}__", keyword)
}