Module asciii::util

Utility functions that are needed all over the places.

Modules

yaml
Yaml Utility functions.

Functions

currency_to_string
One place to decide how to display currency

delete_file_if
Deletes the file if the passed in closure returns true

exit
Exits with the exit status of a child process.

freeze
Freezes the program until for inspection

get_valid_path
takes a path that could be relative or contains a ~ and turn it into a path that exists

git_user_name

ls
Shells out to print directory structure

naive_time_from_str
Creates a chrono::NaiveTime from a string that looks like 23:59:58 or only 12:05.

pass_to_command
Opens the passed paths in the editor set int config.

really
Asks for confirmation

replace_home_tilde
Replaces ~ with $HOME, rust stdlib doesn't do this yet.

setup_log
Sets up logging initially.

to_currency
Creates a currecny from an f64

to_local_file
Changes the extension of a given Path