Crate asciii

Thank you for choosing asciii, the new and improved ascii.

For pure user documentation please refer to the user manual. This contains user and developer documentation together.

Modules

config
Handles config files and default config.

manual
The official asciii Handbook.

print
All the printing code lives here.

project
Project file parsing and evaluation.

storage
Manages file structure of templates, working directory and archives.

templater
Simple templating functionality through keyword replacement.

util
Utility functions that are needed all over the places.

Structs

CONFIG

Functions

version
Returns library version