Module asciii::actions

General actions

Modules

error

Functions

archive_project

csv

delete_project_confirmation

projects_to_csv
Produces a csv string from a list of Projects TODO this still contains german terms

projects_to_tex
Creates the latex files within each projects directory, either for Invoice or Offer.

setup_luigi
Sets up an instance of Storage.

setup_luigi_with_git
Sets up an instance of Storage, with git turned on.

simple_with_projects

spec
Testing only, tries to run complete spec on all projects. TODO make this not panic :D TODO move this to spec::all_the_things

unarchive_projects
Command UNARCHIVE TODO: return a list of files that have to be updated in git

with_projects
Helper method that passes projects matching the search_terms to the passt closure f