Module asciii::project::spec

Implements the ascii invoicer project file specification.

This does all of the heavy lifting. The implementation is separated into sub-modules which take care of separate objectives. Most of the functions in these modules take the yaml data directly as reference. Each module contains a validate() function which ought to be kept up to date.

Modules

archive
Stage 3: requirements to archive

billing
Generates Bills for offer and invoice.

client
Everything about the client

date
All kinds of dates

hours
Everything related to service hours of a project

invoice
Stage 2: requirements for an invoice

offer
Stage 1: requirements for an offer

project
Stage 0: the project itself

Structs

VirtualFieldNames

VirtualFields

Enums

VirtualField
Project::get() allows accessing fields within the raw yaml data structure. Virtual fields are fields that are not present in the document but computed.

Functions

to_currency

Type Definitions

SpecResult