Module asciii::storage

Manages file structure of templates, working directory and archives.

This module takes care of project file management.

Your ordinary file structure would look something like this:

# root dir
├── working
│   └── Project1
│       └── Project1.yml
├── archive
│   ├── 2013
│   └── 2014
│       └── R036_Project3
│           ├── Project3.yml
│           └── R036 Project3 2014-10-08.tex
...

Reexports

pub use self::error::StorageError;

pub use self::storable::Storable;

Modules

error
Error that may occur in Storage

repo

storable
Contains the Storable trait that storable projects must implement.

Structs

ProjectList
Wrapper around Vec<Storable>

Storage
Manages project file storage.

Enums

StorageDir
Used to identify what directory you are talking about.

Functions

list_path_content
Basically ls, returns a list of paths.

Type Definitions

StorageResult
Result returned by Storage

Year
Year = i32