Struct asciii::storage::Storage

pub struct Storage<L: Storable> { /* fields omitted */ }

Manages project file storage.

This includes:

keeping current projects in a working directory

listing project folders and files

listing templates

archiving and unarchiving projects

git interaction

Methods

impl<L: Storable> Storage<L>

pub fn try_new<P: AsRef<Path>>(
root: P,
working: &str,
archive: &str,
template: &str
) -> Result<Self, Error>

Inits storage, does not check existence, yet. TODO

pub fn try_new_with_git<P: AsRef<Path>>(
root: P,
working: &str,
archive: &str,
template: &str
) -> Result<Self, Error>

Inits storage with git capabilities.

pub fn health_check(&self) -> Result<(), Error>

Checks whether the folder structure is as it's supposed to be.

pub fn root_dir(&self) -> &Path

Getter for Storage::storage.

pub fn working_dir(&self) -> &Path

Getter for Storage::working.

pub fn archive_dir(&self) -> &Path

Getter for Storage::archive.

pub fn templates_dir(&self) -> &Path

Getter for Storage::templates.

pub fn extras_dir(&self) -> &Path

Getter for Storage::extras.

pub fn repository(&self) -> Option<&Repository>

Getter for Storage::templates.

pub fn get_repository(&self) -> Result<&Repository, Error>

Getter for Storage::templates, returns Result.

pub fn paths(&self) -> Paths

Returns a struct containing all configured paths of this Storage.

pub fn create_dirs(&self) -> Result<(), Error>

Creates the basic dir structure inside the storage directory.

└── root
    ├── archive
    ├── templates
    └── working

If the directories already exist as expected, that's fine TODO ought to fail when storage_dir already contains directories that do not correspond with the names given in this setup.

pub fn create_archive(&self, year: Year) -> Result<PathBuf, Error>

Creates an archive for a certain year. This is a subdirectory under the archive directory.

└── root
    ├── archive
        ├── 2001
    ...

pub fn list_extra_files(&self) -> Result<Vec<PathBuf>, Error>

Produces a list of files in the extras_dir()

pub fn get_extra_file(&self, name: &str) -> Result<PathBuf, Error>

Returns the Path to the extra file by the given name, maybe.

pub fn list_template_files(&self) -> Result<Vec<PathBuf>, Error>

Produces a list of files in the template_dir()

pub fn list_template_names(&self) -> Result<Vec<String>, Error>

Produces a list of names of all template filess in the templates_dir()

pub fn get_template_file(&self, name: &str) -> Result<PathBuf, Error>

Returns the Path to the template file by the given name, maybe.

pub fn list_archives(&self) -> Result<Vec<PathBuf>, Error>

Produces a list of paths to all archives in the archive_dir. An archive itself is a folder that contains project dirs, therefore it essentially has the same structure as the working_dir, with the difference, that the project folders may be prefixed with the projects index, e.g. an invoice number etc.

pub fn list_years(&self) -> Result<Vec<Year>, Error>

Produces a list of years for which there is an archive.

pub fn create_project(
&self,
project_name: &str,
template_name: &str,
fill_data: &HashMap<&str, String>
) -> Result<L, Error>

Takes a template file and stores it in the working directory, in a new project directory according to it's name.

pub fn archive_project(
&self,
project: &L,
year: Year
) -> Result<Vec<PathBuf>, Error>

Moves a project folder from /working dir to /archive/$year. Also adds the project.prefix() to the folder name.

└── root
    ├── archive
        ├── 2001
            ├── R0815_Birthdayparty
    ...

pub fn archive_projects_if<F>(
&self,
search_terms: &[&str],
manual_year: Option<i32>,
confirm: F
) -> Result<Vec<PathBuf>, Error> where
F: Fn() -> bool,

Moves projects found through search_terms from the Working directory to the Archive/year directory.

Returns list of old and new paths.

pub fn delete_project_if<F>(
&self,
project: &L,
confirmed: F
) -> Result<(), Error> where
F: Fn() -> bool,

pub fn unarchive_projects(
&self,
year: i32,
search_terms: &[&str]
) -> Result<Vec<PathBuf>, Error>

Moves projects found through search_terms from the year back to the Working directory.

Returns list of old and new paths.

pub fn unarchive_project(&self, project: &L) -> Result<PathBuf, Error>

Moves a project folder from /working dir to /archive/$year.

pub fn unarchive_project_dir(
&self,
archived_dir: &Path
) -> Result<PathBuf, Error>

Moves a project folder from /working dir to /archive/$year.

pub fn search_projects(
&self,
directory: StorageDir,
search_term: &str
) -> Result<ProjectList<L>, Error>

Matches StorageDir's content against a term and returns matching project files.

This only searches by name TODO return opened Project, no need to reopen

Warning

Please be adviced that this uses Storage::open_projects() and therefore opens all projects.

pub fn search_projects_any(
&self,
dir: StorageDir,
search_terms: &[&str]
) -> Result<ProjectList<L>, Error>

Matches StorageDir's content against multiple terms and returns matching projects. TODO add search_multiple_projects_deep

pub fn get_project_dir(
&self,
name: &str,
directory: StorageDir
) -> Result<PathBuf, Error>

Tries to find a concrete Project.

pub fn get_project_file(&self, directory: &Path) -> Result<PathBuf, Error>

Locates the project file inside a folder.

This is the first file with the super::PROJECT_FILE_EXTENSION in the folder

pub fn list_project_folders(
&self,
directory: StorageDir
) -> Result<Vec<PathBuf>, Error>

Produces a list of project folders.

pub fn list_empty_project_dirs(
&self,
directory: StorageDir
) -> Result<Vec<PathBuf>, Error>

Produces a list of empty project folders.

pub fn list_project_files(
&self,
directory: StorageDir
) -> Result<Vec<PathBuf>, Error>

Produces a list of project files.

pub fn filter_project_files<F>(
&self,
directory: StorageDir,
filter: F
) -> Result<Vec<PathBuf>, Error> where
F: FnMut(&PathBuf) -> bool,

pub fn open_projects<I>(&self, selection: I) -> Result<ProjectList<L>, Error> where
I: Into<StorageSelection>,

Behaves like list_project_files() but also opens projects directly.

pub fn open_projects_dir(
&self,
directory: StorageDir
) -> Result<ProjectList<L>, Error>

Behaves like list_project_files() but also opens projects directly.

pub fn open_working_dir_projects(&self) -> Result<ProjectList<L>, Error>

pub fn open_all_archived_projects(&self) -> Result<ProjectsByYear<L>, Error>

pub fn open_all_projects(&self) -> Result<Projects<L>, Error>