Struct asciii::storage::Storage

pub struct Storage<L: Storable> {
    pub repository: Option<Repository>,
    // some fields omitted
}

Manages project file storage.

This includes:

keeping current projects in a working directory

listing project folders and files

listing templates

archiving and unarchiving projects

git interaction ( not yet )

Fields

repository

Methods

impl<L: Storable> Storage<L>

fn new<P: AsRef<Path>>(root: P, working: &str, archive: &str, template: &str) -> StorageResult<Self>

Inits storage, does not check existence, yet. TODO

fn new_with_git<P: AsRef<Path>>(root: P, working: &str, archive: &str, template: &str) -> StorageResult<Self>

Inits storage with git capabilities.

fn root_dir(&self) -> &Path

Getter for Storage::storage.

fn working_dir(&self) -> &Path

Getter for Storage::working.

fn archive_dir(&self) -> &Path

Getter for Storage::archive.

fn templates_dir(&self) -> &Path

Getter for Storage::templates.

fn create_dirs(&self) -> StorageResult<()>

Creates the basic dir structure inside the storage directory.

└── root
    ├── archive
    ├── templates
    └── working

If the directories already exist as expected, that's fine TODO ought to fail when storage_dir already contains directories that do not correspond with the names given in this setup.

fn create_archive(&self, year: Year) -> StorageResult<PathBuf>

Creates an archive for a certain year. This is a subdirectory under the archive directory.

└── root
    ├── archive
        ├── 2001
    ...

fn list_template_files(&self) -> StorageResult<Vec<PathBuf>>

Produces a list of files in the template_dir()

fn list_template_names(&self) -> StorageResult<Vec<String>>

Produces a list of names of all template filess in the templates_dir()

fn get_template_file(&self, name: &str) -> StorageResult<PathBuf>

Returns the Path to the template file by the given name, maybe.

fn list_archives(&self) -> StorageResult<Vec<PathBuf>>

Produces a list of paths to all archives in the archive_dir. An archive itself is a folder that contains project dirs, therefore it essentially has the same structure as the working_dir, with the difference, that the project folders may be prefixed with the projects index, e.g. an invoice number etc.

fn list_years(&self) -> StorageResult<Vec<Year>>

Produces a list of years for which there is an archive.

fn create_project(&self, project_name: &str, template_name: &str, fill_data: &HashMap<&str, String>) -> StorageResult<L>

Takes a template file and stores it in the working directory, in a new project directory according to it's name.

fn archive_project_by_name(&self, name: &str, year: Year, prefix: Option<String>) -> StorageResult<PathBuf>

Moves a project folder from /working dir to /archive/$year.

Returns path to new storage dir in archive.

fn archive_project(&self, project: &L, year: Year) -> StorageResult<PathBuf>

Moves a project folder from /working dir to /archive/$year. Also adds the project.prefix() to the folder name.

└── root
    ├── archive
        ├── 2001
            ├── R0815_Birthdayparty
    ...

fn unarchive_project(&self, project: &L) -> StorageResult<PathBuf>

Moves a project folder from /working dir to /archive/$year.

fn unarchive_project_dir(&self, archived_dir: &Path) -> StorageResult<PathBuf>

Moves a project folder from /working dir to /archive/$year.

fn search_projects(&self, directory: StorageDir, search_term: &str) -> StorageResult<Vec<L>>

Matches StorageDir's content against a term and returns matching project files.

This only searches by name TODO return opened Project, no need to reopen

fn search_multiple_projects(&self, dir: StorageDir, search_terms: &[&str]) -> StorageResult<Vec<L>>

Matches StorageDir's content against multiple terms and returns matching project files. TODO add search_multiple_projects_deep

fn get_project_dir(&self, name: &str, directory: StorageDir) -> StorageResult<PathBuf>

Tries to find a concrete Project.

fn get_project_file(&self, directory: &Path) -> StorageResult<PathBuf>

Locates the project file inside a folder.

This is the first file with the super::PROJECT_FILE_EXTENSION in the folder

fn list_project_folders(&self, directory: StorageDir) -> StorageResult<Vec<PathBuf>>

Produces a list of project folders.

fn list_empty_project_dirs(&self, directory: StorageDir) -> StorageResult<Vec<PathBuf>>

Produces a list of empty project folders.

fn list_project_files(&self, directory: StorageDir) -> StorageResult<Vec<PathBuf>>

Produces a list of project files.

fn filter_project_files<F>(&self, directory: StorageDir, filter: F) -> StorageResult<Vec<PathBuf>> where F: FnMut(&PathBuf) -> bool

fn open_project_files(&self, directory: StorageDir) -> StorageResult<ProjectList<L>>

Behaves like list_project_files() but also opens projects directly.