Module asciii::util

Utility functions that are needed all over the places.

Modules

yaml
Yaml Utility functions.

Functions

exit
Exits with the exit status of a child process.

freeze
Freezes the program until for inspection

get_storage_path
Interprets storage path from config.

ls
Shells out to print directory structure

open_in_editor
Opens the passed paths in the editor set int config.

replace_home_tilde
Replaces ~ with $HOME, rust stdlib doesn't do this yet.