Module asciii::util::yaml

Yaml Utility functions.

Here are some missing batteries form the yaml-rust crate. The cool thing about this is the simple path like access to nested elements. if the yaml looks like this:

programmer:
  name: Hendrik
  looks: good
  languages:
    * rust
    * ruby
    * python

you can access "ruby" like this: get_string("programmer/languages/1"). Leading / will not be regarded.

Enums

Yaml
A YAML node is stored as this Yaml enumeration, which provides an easy way to access your YAML document.

YamlError
Wrapper around io::Error and yaml_rust::scanner::ScanError.

Functions

get
Wrapper around get_path().

get_as_string
Gets anything as String.

get_bool
Gets a Bool value.

get_dmy
Gets a Date in dd.mm.YYYY format.

get_f64
Gets a Float value.

get_hash
Gets Some(Yaml::Hash) or None.

get_int
Gets an Int value.

get_str
Gets a &str value.

get_string
same as get_str(), but owned.

open
Wrapper that opens and parses a .yml file.

parse
Ruby like API to yaml-rust.

parse_dmy_date
Interprets "25.12.2016" as date.

parse_dmy_date_range
Interprets "24-25.12.2016" as date.